feat: make all result pages consistent
